Tested working as a service. Tray icon is not accessible due to services running in isolation mode (0) where user exe is run in (1) allowing it to appear. Revising to not run program as service. Instead it will embed in program files. Seperated installer from main to remove main_requires_admin
