+++ tenant-client/helperfunctions.py
@@ -192,61 +192,3 @@ def tenant_portal():
     # add check here to ensure starts with https://
     os.system(f"start {TENANT_PORTAL_URL}")
 
-@staticmethod
-def first_run(arg):
-    """
-    The first run function checks if this program has been run before. 
-    If it has not been run before it will create a registry entry and 
-    call the API to send the success install
-
-    :param arg: The download key
-    :return: True if the first run is successful, False otherwise
-    """
-
-    # call API.get_download_key
-    download_key = API.get_download_key()
-    if download_key != arg :
-        return False
-        # create registry entry Computer\HKEY_LOCAL_MACHINE\SOFTWARE\Nexum\A2
-    try:
-        check_install_key("","","",5001) # send install key and tenant secret to verify install
-        get_client_info()
-        save_client_info()
-        key = winreg.CreateKey(winreg.HKEY_LOCAL_MACHINE, r"SOFTWARE\Nexum")
-        winreg.CloseKey(key)
-        key = winreg.CreateKey(winreg.HKEY_LOCAL_MACHINE, r"SOFTWARE\\Nexum\\Client")
-        winreg.CloseKey(key)
-        API.send_success_install(CLIENT_ID,TENANT_ID,CLIENT_SECRET)
-        return True
-    except FileNotFoundError:
-        print(traceback.format_exc())
-        return False
-    except PermissionError:
-        print(traceback.format_exc())
-        return False
-    except:
-        print(traceback.format_exc())
-        return False
-@staticmethod
-def check_first_run(arg):
-
-    """
-    The check_first_run function checks if this program has been run before.
-    It is a helper function for the first_run function
-    """
-    l = Logger()
-    # check for registry entry Computer\HKEY_LOCAL_MACHINE\SOFTWARE\Nexum\A2
-    # to see if this is the first run if it exists return, else call first_run()
-    try:
-        key = winreg.OpenKey(winreg.HKEY_LOCAL_MACHINE, r"SOFTWARE\\Nexum\\Client")
-        winreg.CloseKey(key)
-        return True
-    except FileNotFoundError:
-        return first_run(arg)
-    except PermissionError:
-        l.log("ERROR", "check_first_run", "Permission Error checking registry",
-        "1005", time.strftime("%Y-%m-%d %H:%M:%S:%m", time.localtime()))
-    except:
-        l.log("ERROR", "check_first_run", "General Error checking registry",
-        "1002", time.strftime("%Y-%m-%d %H:%M:%S:%m", time.localtime()))
-    return False
