+++ tenant-client/main.py
@@ -40,33 +40,27 @@ Error Codes
 1003 - File not found
 1004 - Decryption failed
 1005 - Permission error
+
 404 - Not found
 401 - Access denied
-500 - Internal server error
+503 - Internal server error
 """
 
 # pylint: disable= no-member,no-name-in-module, import-error
 
-import win32serviceutil
-import win32service 
-import win32event
-import servicemanager
-import socket
-import sys
-import threading
+
 import time
-from pyuac import main_requires_admin
 from logger import Logger
 from initsql import InitSql
 from runjob import RunJob, LOCAL_JOB
-from helperfunctions import get_client_info, logs, tenant_portal,check_first_run
+from helperfunctions import get_client_info, logs, tenant_portal
 from security import Security
 from jobsettings import JobSettings
 from iconmanager import IconManager, image_path
 from flaskserver import FlaskServer
 # Global variables
 
-@main_requires_admin
+
 def main():
     """
     Main method of the program for testing and starting the program
@@ -78,8 +72,6 @@ def main():
     t.password = "Test123"
     LOCAL_JOB.set_settings(t)
     Security.set_client_secret("ASDFGLKJHTQWERTYUIOPLKJHGFVBNMCD")
-    # check if this is the first run
-    check_first_run("1234")
     # create a Logger
     l = Logger()
     # init databases
@@ -101,43 +93,11 @@ def main():
 
     # run server to listen for requests
     FlaskServer()
+    while True:
+        pass
 
-class TestService(win32serviceutil.ServiceFramework):
-    _svc_name_ = "NexumClientService"
-    _svc_display_name_ = "Nexum Client"
 
-    def __init__(self, args):
-        win32serviceutil.ServiceFramework.__init__(self, args)
-        self.hWaitStop = win32event.CreateEvent(None, 0, 0, None)
-        self.is_alive = True
 
-    def SvcStop(self):
-        self.ReportServiceStatus(win32service.SERVICE_STOP_PENDING)
-        win32event.SetEvent(self.hWaitStop)
-        self.is_alive = False
 
-    def SvcDoRun(self):
-        servicemanager.LogMsg(servicemanager.EVENTLOG_INFORMATION_TYPE,
-                              servicemanager.PYS_SERVICE_STARTED,
-                              (self._svc_name_, ''))
-        self.main()
-    
-
-    def main(self):
-        t1 = threading.Thread(target=main)
-        t1.setDaemon(True)
-        t1.start()
-        win32event.WaitForSingleObject(self.hWaitStop, win32event.INFINITE)
-        while True:
-            pass
-
-if __name__ == "__main__":
+if __name__ == "__main__": 
     main()
-    if len(sys.argv) > 1:
-       # Called by Windows shell. Handling arguments such as: Install, Remove, etc.
-       win32serviceutil.HandleCommandLine(TestService)
-    else:
-       # Called by Windows Service. Initialize the service to communicate with the system operator
-       servicemanager.Initialize()
-       servicemanager.PrepareToHostSingle(TestService)
-       servicemanager.StartServiceCtrlDispatcher()
